Fixes incomplete reception buffer flushing (ticket #55)
There is a race condition between the time a socket is created and the
time it is initialized (BPF filter pushed, bound to an interface). Indeed,
packets captured in between would still be delivered when reading the socket.
Setting reception buffers to 0 was part of the solution and greatly reduced
the race condition timeframe, but still, packet already captured where actually
not sweeped out. So this also need to be done. Thanks to AGR and Y.E. Jutard.