Support for special 'Scapy' variant of pypcap to enable timeouts
pylibpcap and pcapy offer a 'to_ms' parameter to set the read
timeout in milliseconds, but standard pypcap does not.
(Instead there is the 'immediate' parameter to disable buffering.)
A specially patched pcap version was made for Scapy which adds a
timeout_ms parameter, but to stay compatible with pypcap it has a
different position in the argument list than in pylibpcap
and pcapy.
So we have to distinguish four different Python bindings for
libpcap/WinPcap.